A simple tutorial of calling CPLEX/Gurobi/Mosek in Julia
A quick-start tutorial on how to call the state-of-the-art optimisation solvers, such as IBM ILOG CPLEX Optimizer, Gurobi, and Mosek, through the Julia programming lauguage. Note that the purpose of this tutorial is educational, instead of jeopardizing the credits of the official documents of JuMP. Actually, many code snippets are from the JuMP official documents. This tutorial enables you to start formulating and solving your models via CPLEX/Gurobi/Mosek within 2 hours.